updating notificationViewModel to use geofencing and notification feature
